Add support for iCloud Keychain
getting password will tell you if it's synchronizable, and when you set a password you can decide whether to store it locally or in iCloud. Marking something synchronizable works fine even if you don't have iCloud Keychain or aren't signed into iCloud.

Also added the ability to set a label for the keychain item (passing in nil will set it to the same as the service name). This is because items in iCloud Keychain are visible from a Mac using Keychain Access